prélèvement ressources et corrections

+++ src/solar_system/Player.java
@@ -20,7 +20,7 @@ public class Player {
 	
 
 	
-	public void addRessource(String type , int augmentation ){
+	public void addRessource(String type , double augmentation ){
 		if(ressources.containsValue(type)){
 		ressources.put(type,ressources.get(type)+augmentation);	
 		// Fabien nous garantit que ca ne va pas merder.
